refactor:apiDirectoryUpdate #34
